Fix auth redirects to respect the return URL.
When the user refreshes the screen or authentication token is
invalid, the redirect URL was incorrect and always defaulting to
the root page.

This change fixes a bug where it was using `this.return` instead
of `this.returnUrl` and fixes the AuthGuard to actually put back
the URL state when redirecting.