Redirect the user back to Login page if unauthorized
Angular HTTP Client doesn't let you issue another request when
requesting the next URL unless you add it to the queue, which is
not what we are doing. So instead, if the service is coming back
as 401, we can redirect the user back automatically to the login
page with the return URL in tact.

The return URL is needed to create a positive experience instead
navigating back where they are coming from after logging back in.

Fixes #109
Part of #116