Persist Authentication Token in Session Storage
When refreshing the page, the token is lost since it was saved
in memory. By using session storage, we can ensure that browser
session will keep the token there until they close the tab.

In the future we can move this to localStorage if we have a way
to validate the token expiration and renew the token.

For #116