Update ethers.js to 5.3.0
Ethers.js prior to 5.3.0 was using a vulnerable version of ws (7.2.3) which is vulnerable to Regular Expression Denial of Service (ReDoS). A specially crafted value of the Sec-Websocket-Protocol header can be used to significantly slow down a ws server. The vulnerability has been fixed in ws@7.4.6 which has been implemented in ethers.js 5.3.0.